requirements.txt
flask
tensorflow
pillow
keras
opencv-python
matplotlib
pandas
